去除gtest依赖
