+++ .github/workflows/linux.yml
@@ -19,7 +19,7 @@ jobs:
     - uses: actions/checkout@v3
     - run: sudo apt update
     - run: sudo apt install build-essential cmake ninja-build
-    - run: sudo apt install libjsoncpp-dev libgflags-dev libgtest-dev libgoogle-glog-dev libunwind-dev
+    - run: sudo apt install libjsoncpp-dev libgflags-dev libgoogle-glog-dev libunwind-dev
 
     - name: Setup Ninja
       uses: ashutoshvarma/setup-ninja@master
