+++ templates/CMakeLists.txt
@@ -14,7 +14,6 @@ SET(MYFRAME_INSTALL_LIB_DIR ${CMAKE_INSTALL_PREFIX}/lib)
 #### dep 3rd lib
 FIND_PACKAGE(jsoncpp REQUIRED)
 FIND_PACKAGE(gflags REQUIRED)
-FIND_PACKAGE(GTest REQUIRED)
 
 #### include directory
 INCLUDE_DIRECTORIES(${CMAKE_INSTALL_PREFIX}/include)
